Some improvements from review
